requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.2
sqlparse==0.5.0
urllib3==2.2.1
whitenoise==6.6.0
